# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OCC Operational Control Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List complaints
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: ptId
          in: query
          schema:
            type: integer
        - name: severity
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Complaint list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplaintWithRelations"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Create complaint
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplaintRequest"
      responses:
        "201":
          description: Complaint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintWithRelations"
  /complaints/{id}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get complaint details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Complaint details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintWithRelations"
    put:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplaintRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintWithRelations"
...
  /chats:
    get:
      operationId: listChats
      tags: [chats]
      summary: List user chats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Chat list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatWithMembers"
    post:
      operationId: createChat
      tags: [chats]
      summary: Create chat room
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatRequest"
      responses:
        "201":
          description: Chat created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatWithMembers"
  /chats/{id}/members:
    post:
      operationId: addChatMember
      tags: [chats]
      summary: Add member to chat
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
              required: [userId]
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /chats/{id}/messages:
    get:
      operationId: listChatMessages
      tags: [chats]
      summary: List chat messages
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chats]
      summary: Send chat message
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageRequest"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /handover-logs:
    get:
      operationId: listHandoverLogs
      tags: [handover]
      summary: List handover logs
      security:
        - bearerAuth: []
      parameters:
        - name: ptId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Handover logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HandoverLogWithRelations"
    post:
      operationId: createHandoverLog
      tags: [handover]
      summary: Create handover log
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHandoverLogRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HandoverLogWithRelations"
  /handover-logs/{id}:
    get:
      operationId: getHandoverLog
      tags: [handover]
      summary: Get handover log by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Handover log details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HandoverLogWithRelations"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    ComplaintWithRelations:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        complaintType:
          type: string
        ptId:
          type: integer
        branchId:
          type: integer
        assignedUserId:
          type: integer
        assignedUserName:
          type: string
        severity:
          type: string
        chronology:
          type: string
        followUp:
          type: string
        status:
          type: string
        createdBy:
          type: integer
        creatorName:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, complaintType, severity, status]
    CreateComplaintRequest:
      type: object
      properties:
        title:
          type: string
        complaintType:
          type: string
        ptId:
          type: integer
        branchId:
          type: integer
        assignedUserId:
          type: integer
        severity:
          type: string
        chronology:
          type: string
      required: [title, complaintType, severity]
    UpdateComplaintRequest:
      type: object
      properties:
        title:
          type: string
        severity:
          type: string
        status:
          type: string
        assignedUserId:
          type: integer
        followUp:
          type: string
...
    ChatWithMembers:
      type: object
      properties:
        id:
          type: integer
        chatType:
          type: string
        name:
          type: string
        createdBy:
          type: integer
        members:
          type: array
          items:
            $ref: "#/components/schemas/ChatMember"
        lastMessage:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, chatType]
    ChatMember:
      type: object
      properties:
        userId:
          type: integer
        userName:
          type: string
        joinedAt:
          type: string
          format: date-time
      required: [userId]
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        chatId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        message:
          type: string
        attachmentUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, chatId, senderId, message]
    SendChatMessageRequest:
      type: object
      properties:
        message:
          type: string
        attachmentUrl:
          type: string
      required: [message]
    CreateChatRequest:
      type: object
      properties:
        chatType:
          type: string
        name:
          type: string
        memberIds:
          type: array
          items:
            type: integer
      required: [chatType, memberIds]
    HandoverLogWithRelations:
      type: object
      properties:
        id:
          type: integer
        ptId:
          type: integer
        ptName:
          type: string
        branchId:
          type: integer
        fromShiftId:
          type: integer
        fromShiftName:
          type: string
        toShiftId:
          type: integer
        toShiftName:
          type: string
        createdBy:
          type: integer
        creatorName:
          type: string
        summary:
          type: string
        pendingActivities:
          type: string
        pendingTasks:
          type: string
        pendingComplaints:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id]
    CreateHandoverLogRequest:
      type: object
      properties:
        ptId:
          type: integer
        branchId:
          type: integer
        fromShiftId:
          type: integer
        toShiftId:
          type: integer
        summary:
          type: string
        pendingActivities:
          type: string
        pendingTasks:
          type: string
        pendingComplaints:
          type: string
        notes:
          type: string
      required: [fromShiftId, toShiftId]
...